Move all configuration to config.js
