+++ bin/simple-subnet-api.js
@@ -1,12 +1,6 @@
 import '../lib/instrument.js'
 import { createApp } from '../lib/app.js'
-import { DATABASE_URL } from '../lib/migrate.js'
-
-const {
-  PORT = '8080',
-  HOST = '127.0.0.1',
-  REQUEST_LOGGING = 'true'
-} = process.env
+import { DATABASE_URL, HOST, PORT, REQUEST_LOGGING } from '../lib/config.js'
 
 const app = createApp({
   databaseUrl: DATABASE_URL,
